# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ToDoToDoToDo API",
    "description": "할일 관리 애플리케이션 REST API - 사용자별 할일 추가, 조회, 수정, 삭제, 완료 처리, 복구, 우선순위 변경 기능 제공",
    "version": "1.0.0",
    "contact": {
      "name": "ToDoToDoToDo Team"
    }
  },
  "servers": [
    {
      "url": "https://api.todotodotodo.com",
      "description": "프로덕션 서버"
    },
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 토큰 기반 인증 (24시간 유효)"
      }
    },
    "schemas": {
      "SignUpRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "사용자 이메일 (RFC 5322 형식, 중복 불가)",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "pattern": "^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>\\/?]).*$",
            "description": "비밀번호 (8자 이상, 영문/숫자/특수문자 조합 필수)",
            "example": "Password123!"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "사용자 이름 (2-50자)",
            "example": "홍길동"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "사용자 이메일",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "Password123!"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT 액세스 토큰 (24시간 유효)",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string",
                "format": "uuid",
                "description": "사용자 고유 ID",
                "example": "550e8400-e29b-41d4-a716-446655440000"
              },
              "email": {
                "type": "string",
                "format": "email",
                "description": "사용자 이메일",
                "example": "user@example.com"
              },
              "name": {
                "type": "string",
                "description": "사용자 이름",
                "example": "홍길동"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "유효하지 않은 입력입니다"
          },
          "details": {
            "type": "object",
            "description": "에러 상세 정보 (선택사항)",
            "example": {
              "field": "startDate",
              "reason": "시작일은 종료일 이전이거나 같아야 합니다"
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": ["인증"],
        "summary": "회원가입",
        "description": "새로운 사용자 계정을 생성합니다",
        "operationId": "signUp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 이메일, 비밀번호 규칙 위반, 중복 이메일 등)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "이미 사용 중인 이메일입니다",
                  "details": {
                    "field": "email"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["인증"],
        "summary": "로그인",
        "description": "이메일과 비밀번호로 로그인하여 JWT 토큰을 받습니다",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패 (잘못된 이메일 또는 비밀번호)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "이메일 또는 비밀번호가 올바르지 않습니다"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}